requirements.txt
# requirements.txt

PyYAML>=5.3.1
numpy>=1.18.0
pandas>=1.0.0
scikit-learn>=0.22.0
imbalanced-learn>=0.6.0
pynvml
tqdm>=4.41.0
matplotlib>=3.1.0
opencv-python>=4.1.0
albumentations>=0.4.0
Pillow>=7.0.0
ultralytics
torchmetrics>=1.0.0,<2.0.0
scikit-plot>=0.3.7
